<refactor>: 기록 누적 페이지 생성 및 파일명 snake_case 적용
